Serialize data in a language agnostic manner
The app was publishing messages with serialized internal data
structures.  To make it more interoperable with clients/subscribers
that don't have access to these data structures or are written in another
programming language, I made the app to serialize/deserialize the data
to the JSON format with help of Niels Lohmann's library.